requirements.txt
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.24.0
pandas_ta>=0.3.14b1
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
vaderSentiment>=3.3.2
transformers>=4.35.0
torch>=2.1.0
rich>=13.6.0
lxml>=4.9.0
html5lib>=1.1
urllib3>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.39.0
openai>=1.50.0
